implement Transaction mod

+++ dev/backend/src/lib.rs
@@ -36,7 +36,7 @@ async fn create() -> Result<CreateResponse, String> {
 }
 
 #[update]
-async fn sign_evm_tx(hex_raw_tx: Vec<u8>, chain_id: u8) -> Result<SignatureInfo, String> {
+async fn sign_evm_tx(hex_raw_tx: Vec<u8>, chain_id: usize) -> Result<SignatureInfo, String> {
     let principal_id = ic_cdk::caller();
     let res = no_key_wallet::sign(hex_raw_tx, chain_id, principal_id)
         .await
